+++ packages/solid/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Improve type of submit event at `Form` and `handleSubmit` (issue #18)
+- Change build step from tsup to Rollup and update exports (issue #16)
 
 ## v0.9.5 (January 07, 2023)